update encode varint to vec (#20)
* update encode varint to vec

* update RunesId's block to u64 instead of u32

* update RunesId's block to u64 instead of u32

+++ customs/bitcoin/src/state.rs
@@ -67,7 +67,7 @@ impl Error for ParseRuneIdError {
     Deserialize,
 )]
 pub struct RuneId {
-    pub block: u32,
+    pub block: u64,
     pub tx: u32,
 }
